requirements.txt
django
requests
django-rest-framework
bs4
django-filter
python-dateutil
feedparser
pytest
pytest-django